+++ spec/lib/roadie/rails/railtie_spec.rb
@@ -42,6 +42,19 @@ module Roadie
           expect(providers[1]).to be_instance_of(AssetPipelineProvider)
           expect(providers[1].pipeline).to eq(asset_pipeline)
         end
+
+        # This happens inside a Rails engine as the parent app is the one
+        # holding on to the pipeline.
+        it "gets no AssetPipelineProvider if assets are enabled but not available" do
+          rails_application.config.assets = ActiveSupport::OrderedOptions.new(enabled: true)
+          allow(rails_application).to receive(:assets).and_return nil
+
+          run_initializer
+
+          providers = Railtie.config.roadie.asset_providers.to_a
+          expect(providers).to have(1).item
+          expect(providers[0]).to be_instance_of(FilesystemProvider)
+        end
       end
     end
   end